SESSION FLOW - Similar requests: leave a float number with 2 decimal digits & leave the % after the number(continuation)

+++ js/telepath.sessionflow.js
@@ -100,8 +100,8 @@ telepath.sessionflow = {
 					if(similarity.username){
 						table.append(getRow('User:', similarity.username));
 					}
-					
-					table.append(getRow('Similarity:',parseInt(similarity._score).toFixed(2)+ '%'));
+
+					table.append(getRow('Similarity:', parseFloat(similarity._score).toFixed(2) + '%'));
 					
 					//if(alert.user && alert.user != '') {
 					//	table.append(getRow('User:', alert.user));
